Automated Update Notice Issue
The automated updated notice does not appear for the addon on the
wordpress->Plugins page. This was found when we tried releasing version
1.4. The reason for the issue is that the addon uses the option name for
the 'Printable Tickets Addon' when searching for the license key. As a
result, no match is found and inspite of having an active license, the
update notification does not appear.

This fix will be released in version 1.4