WPML Multi-currency Compatibility Fix
Scenario: Site uses WPML, However, the multi currency setting is
disabled. In such a scenario, the prices are not displayed on the
Product page.  This is because the formatting from the WPML filters
fail. Hence we have now added a condition to use the WPML filters only
if multi currency is enabled.