Version 1.5
Multiple Time Slots Addon: Changelog v 1.5

Please note that the addon page should be upgraded to mention that the
Multiple Time Slots Addon v1.5 will be compatible with Booking 2.4.4 and
above.

This upgrade to the addon contains some issue fixes as well as
compatibility fixes.

1. The addon has now been modified to make it compatible with the price
calculation changes done for the WooCommerce Booking and Appointment
plugin (v2.4.4 ).
2. The issue where the time slot prices were not displayed on the
Product page ( for a site using WPML, but the multi-currency feature is
disabled ) has now been fixed.