WPML Multicurrency Compatibility Fix
Time slot prices were calculated based on the exchange rates price only,
even if the "Set prices in other currencies manually" option is selected
on the admin products page.

The Time slots should be calculated on the manual prices set for "Set
prices in other currencies manually" instead of the exchanged prices for
the products.