The currency is coming wrong from the plugin when wpml multi-currency is enabled.
Scenario: When multi-currency is set on the product level and when a
date is selected, the price is coming right from the plugin, but the
currency symbol is coming wrong.

When the selected currency is Dirhams, it is displaying symbol for Euros
(which is the default currency set from WooCommerce).

http://screencast.com/t/rN39nsau