Remove wcml_formatted_price filter
Remove the wcml_formatted_price filter.

Reasons:
1. In the last update we've added code to ensure that when the currency
is set to manual for a product from WCML plugin the currency symbol
should be correct.
2. In the latest interactions with the WPML support team, we have been
updated that the wcml_formatted_price filter is no longer supported.

It anyways becomes repetitive when we're setting the currency manually.
The same patch can be used for all types of scenarios:
1. WPML is not being used
2. WPML is used, but multicurrency is not being used
3. WPML & Multi currency are being used (automated)
4. WPML & multi currency are being used (manual).

Hence, it's best to remove the filter and use the new code.