fix #18
The issue was occurring because we were passing the base price itself
instead of multiplying it with quantity. In this fix, we are calculating
the price as per the quantity so is the Multiple Time Slot Addon is
active and product is only setup with the single day booking method then
price will be calculated correct as per the selected quantity.